Like button toggle and URL cleanup
Added functionality to unlike a song when clicking the heart button. Changed around the js and html for the heart link to make it work a bit better. Added window.history.replaceState to the script to cleanup URLs after the user logs in. Also changed the heart link to prevent reloading/# symbols displayed in URL.

May need to test the reload functionality in other browsers.